Clarify and rename identity category
Per #110 , proposing initial changes to clarify the Identity category, and rename it to Identity Lifecycle and Access Management. 

Hopefully this makes it clear that it's not just about each service's own entitlements or authn/authz libraries, but really focused on tools which make it easier for all workloads to support authentication, secure use of tokens, manageable entitlements and control the lifecycle and auditability of those.