Small tweaks on the README
"List of Meetings" is redundant in the "Meeting times" because
it duplicates the information in the "Additional Information" section.
Another thing is that when you click it, it doesn't lead you to
the right place.

Member list should be fixed too - it too doesn't lead you to the right
section on the README file.

Signed-off-by: Martin Vrachev <mvrachev@vmware.com>