Update ci/spelling-config.json
Comment from @JohnHillegass: So I think where this is messing up is we excluded that whole .github dir from cspell execution but grep is seeing markdown files changed there and passing forward to xargs. Since the first arg is not empty xargs is taking it and passing to cspell. Cspell since its been told to skip that path, returns a non-zero code and xargs throws the 123. Changing like above handles this case.

Co-authored-by: John Hillegass <JohnHillegass@users.noreply.github.com>