Update README.md
Update meeting ID. 

The motivation for this change, automated livestream and upload to youtube broke a while back with changes to how livestream is configured. It no longer allows use of streaming using personal room id.

Also retitled US and EMEA to US and Western Hemisphere and Eastern Hemisphere for inclusivity.

Signed-off-by: Andres Vega <av@monkey.org>