Change in Designated Facilitator Role for TAG Security Assessments
Stepping down from my position as one of the designated facilitator for the TAG security assessments after evaluating my current commitments and responsibilities, both professionally and personally.

The process is currently very effective with the time and energy that Justin dedicates to it, and I'm eager to see how the TAG security assessments continue to grow and evolve.

I look forward to continuing to contribute to the TAG in the capacity of TL.

Signed-off-by: Andres Vega <av@monkey.org>
Signed-off-by: Mariusz Sabath <mrsabath@gmail.com>